feat(FN-3575): move form values and errors VM into main file
