--- trade-finance-manager-ui/server/types/view-models/record-correction/create-record-correction-request-errors-view-model.ts
@@ -1,7 +1,0 @@
-import { ErrorSummaryViewModel } from '../error-summary-view-model';
-
-export type CreateRecordCorrectionRequestErrorsViewModel = {
-  reasonsErrorMessage?: string;
-  additionalInfoErrorMessage?: string;
-  errorSummary: ErrorSummaryViewModel[];
-};
