--- trade-finance-manager-ui/server/types/view-models/record-correction/create-record-correction-request-form-values.ts
@@ -1,6 +1,0 @@
-import { RecordCorrectionRequestReason } from '@ukef/dtfs2-common';
-
-export type CreateRecordCorrectionRequestFormValues = {
-  reasons?: RecordCorrectionRequestReason[];
-  additionalInfo?: string;
-};
