+++ trade-finance-manager-ui/server/types/view-models/record-correction/index.ts
@@ -1,3 +1,1 @@
 export * from './create-record-correction-request-view-model';
-export * from './create-record-correction-request-errors-view-model';
-export * from './create-record-correction-request-form-values';
